Consolidate added dependency inference tests
